Realizo ajustes en la sección Elements - RadioButtons/Buttons/TextBox
